+++ pyproject.toml
@@ -1,3 +1,15 @@
+[build-system]
+requires = ["setuptools>=61"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "otxlearner"
+version = "0.1.0"
+description = "Sinkhorn-penalised X-Net for causal inference"
+readme = "README.md"
+requires-python = ">=3.10"
+license = {file = "LICENSE"}
+
 [tool.black]
 line-length = 88
 
@@ -7,4 +19,4 @@ line-length = 88
 [tool.mypy]
 python_version = "3.10"
 strict = true
-packages = ["src"]
+packages = ["otxlearner"]
