+++ tests/test_acic_loader.py
@@ -2,10 +2,8 @@ from __future__ import annotations
 
 import numpy as np
 from pathlib import Path
-import sys
 
-sys.path.append(str(Path(__file__).resolve().parents[1]))
-from src.data import load_acic
+from otxlearner.data import load_acic
 
 
 def _create_fake_dataset(path: Path) -> None:
