+++ tests/test_gradient_reversal.py
@@ -1,12 +1,7 @@
 from __future__ import annotations
 
-from pathlib import Path
-import sys
-
-sys.path.append(str(Path(__file__).resolve().parents[1]))
-
 import torch
-from src.models.domain import GradientReversal
+from otxlearner.models.domain import GradientReversal
 
 
 def test_gradient_reversal_backward() -> None:
