+++ tests/test_mlp.py
@@ -1,11 +1,6 @@
 from __future__ import annotations
 
-import sys
-from pathlib import Path
-
-sys.path.append(str(Path(__file__).resolve().parents[1]))
-
-from src import MLPEncoder
+from otxlearner import MLPEncoder
 import torch
