+++ tests/test_twins_loader.py
@@ -2,10 +2,7 @@ from __future__ import annotations
 
 import numpy as np
 from pathlib import Path
-import sys
-
-sys.path.append(str(Path(__file__).resolve().parents[1]))
-from src.data import load_twins
+from otxlearner.data import load_twins
 
 
 def _create_fake_dataset(path: Path) -> None:
